Adopts sverchoks node structure
