+++ nodes/number.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  make_range.py
+#  number.py
 #
 #  Copyright (c) 2022 Ronny Scharf-Wildenhain <ronny.scharf08@gmail.com>
 #
@@ -22,43 +22,40 @@
 #
 #
 ###################################################################################
-import os
-import awkward as ak
-import numpy as np
+from decimal import Decimal
+
+from qtpy.QtWidgets import QLineEdit
+from nodeeditor.node_content_widget import QDMNodeContentWidget
 
 from fcn_conf import register_node
-from fcn_base_node import FCNNode
+from fcn_base_node import FCNNode, FCNNodeContentView
 from fcn_locator import icon
 
 
 @register_node
-class MakeRange(FCNNode):
+class Number(FCNNode):
 
     icon: str = icon("fcn_default.png")
-    op_title: str = "Make Range"
-    op_category = "Math"
+    op_title: str = "Number"
+    op_category = "Number"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
         super().__init__(scene=scene,
-                         inputs_init_list=[(0, "Start", 1, 0, True, ('int', 'float')),
-                                           (0, "Stop", 1, 10, True, ('int', 'float')),
-                                           (0, "Step", 1, 1, False, ('int', 'float'))],
+                         inputs_init_list=[(0, "In", 1, 0, False, ('int', 'float'))],
                          outputs_init_list=[(0, "Out", 0, 0, True, ('int', 'float'))],
                          width=150)
 
+    def collapse_node(self, collapse: bool = False):
+        super().collapse_node(collapse)
+
+        if (collapse is True) and isinstance(self.sockets_input_data[0][0], (int, float)):
+            self.title: str = 'In: %.2E' % Decimal(str(self.sockets_input_data[0][0]))
+        else:
+            self.title: str = self.default_title
+
     def eval_operation(self, sockets_input_data: list) -> list:
-        # Inputs
-        start = sockets_input_data[0]
-        stop = sockets_input_data[1]
-        step = sockets_input_data[2][0]
+        self.collapse_node(self.content.isHidden())
 
-        # Force array broadcast
-        start, stop = ak.broadcast_arrays(start, stop)
-
-        res = []
-        for idx, _start in enumerate(ak.flatten(start, axis=None)):
-            _stop = ak.flatten(stop, axis=None)[idx]
-            res.append(np.arange(_start, _stop, step).tolist())
-
-        return [res]
+        in_val: float = float(sockets_input_data[0][0])
+        return [[in_val]]
