+++ nodes/point.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  make_sphere.py
+#  sphere.py
 #
 #  Copyright (c) 2022 Ronny Scharf-Wildenhain <ronny.scharf08@gmail.com>
 #
@@ -34,11 +34,11 @@ from fcn_utils import flatten_to_tuples
 
 
 @register_node
-class MakePoint(FCNNode):
+class Point(FCNNode):
 
     icon: str = icon("fcn_default.png")
-    op_title: str = "Make Point"
-    op_category = "FreeCAD"
+    op_title: str = "Point"
+    op_category = "Generator"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
