+++ nodes/vector_in.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  basic_math.py
+#  vector_in.py
 #
 #  Copyright (c) 2022 Ronny Scharf-Wildenhain <ronny.scharf08@gmail.com>
 #
@@ -22,6 +22,7 @@
 #
 #
 ###################################################################################
+import os
 import awkward as ak
 
 from fcn_conf import register_node
@@ -30,47 +31,28 @@ from fcn_locator import icon
 
 
 @register_node
-class BasicMath(FCNNode):
+class VectorIn(FCNNode):
 
     icon: str = icon("fcn_default.png")
-    op_title: str = "Basic Math"
-    op_category = "Math"
+    op_title: str = "Vector In"
+    op_category = "Vector"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
         super().__init__(scene=scene,
-                         inputs_init_list=[(0, "Op", 3, ["a+b", "a-b", "a*b", "a/b", "a^b", ], False, ('int', )),
-                                           (0, "a", 1, 1, True, ('float', 'int')),
-                                           (0, "b", 1, 10, True, ('float', 'int'))],
-                         outputs_init_list=[(0, "Res", 0, 11, True, ('float', 'int'))],
+                         inputs_init_list=[(0, "X", 1, 1.0, True, ("int", "float")),
+                                           (0, "Y", 1, 0.0, True, ("int", "float")),
+                                           (0, "Z", 1, 0.0, True, ("int", "float"))],
+                         outputs_init_list=[(1, "Vec", 0, 0, True, ("vec", ))],
                          width=150)
-
-    def collapse_node(self, collapse: bool = False):
-        super().collapse_node(collapse)
-
-        if collapse is True:
-            self.title = "Math: " + self.content.input_widgets[0].currentText()
-        else:
-            self.title = self.default_title
 
     def eval_operation(self, sockets_input_data: list) -> list:
         # Inputs
-        op_code: int = sockets_input_data[0][0]
-        a_array = ak.Array(sockets_input_data[1])
-        b_array = ak.Array(sockets_input_data[2])
+        x_in = sockets_input_data[0]
+        y_in = sockets_input_data[1]
+        z_in = sockets_input_data[2]
 
-        # Outputs
-        if op_code == 0:  # Add
-            res = a_array + b_array
-        elif op_code == 1:  # Sub
-            res = a_array - b_array
-        elif op_code == 2:  # Mul
-            res = a_array * b_array
-        elif op_code == 3:  # Div
-            res = a_array / b_array
-        elif op_code == 4:  # Pow
-            res = a_array ** b_array
-        else:
-            raise ValueError("Unknown operation (Op)")
-
+        # Broadcast an zip to vector
+        x_vector, y_vector, z_vector = ak.broadcast_arrays(x_in, y_in, z_in)
+        res = ak.zip([x_vector, y_vector, z_vector])
         return [res.tolist()]
